urls.py: View function names changed to class based views.
views.py:
-NEW-
Create View Class: NewDogTag.
CreateView - used when you need a form on the page
and need to do a db insertion on submission of a valid form.

This was not required but to better my understanding of Create Views.

-DELETED-
Functional Views that were replaced by a Create View Class:
new_dog_tag.